Add set_env_key to shell interface
Using set_env (which takes a string of the type KEY=VALUE) is not very
handy when the key and value are already separate (for instance, in the
case of CD where the keys are hard-coded).
Set_env is probably gonna disappear, and expression of the type
KEY=VALUE will be cutted beforehand.